src/1 - Extract Features/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
timm>=0.6.0
matplotlib>=3.4.0
scikit-learn>=0.24.0 